python_backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
aiomysql>=0.2.0
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
python-socketio[asyncio]>=5.11.0
python-multipart>=0.0.12
httpx>=0.27.0
openai>=1.57.0
APScheduler>=3.10.4
cryptography>=43.0.0
aiogram>=3.15.0
python-dotenv>=1.0.0
aiofiles>=24.1.0
